pyproject.toml
[project]
name = "voice-agent-trial"
version = "0.1.0"
description = "Real-time voice agent platform using Pipecat, Cartesia, Cerebras, and ElevenLabs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[daily,cartesia,cerebras,elevenlabs,silero]>=0.0.92",
    "python-dotenv>=1.2.1",
    "loguru>=0.7.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = ["ruff>=0.14.3", "ty<=0.0.1a25"]

[project.scripts]
voice-agent = "voice_agent:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["config", "services", "utils"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
